+++ django_test_app/server/apps/token_auth/urls.py
@@ -20,6 +20,21 @@ router = Router(
             name='token_obtain_async',
         ),
         path(
+            'token-default-sync-auth/',
+            example.ControllerWithDefaultTokenSyncAuth.as_view(),
+            name='token_default_sync_auth',
+        ),
+        path(
+            'token-concrete-obtain-sync/',
+            obtain.ConcreteObtainTokenSyncController.as_view(),
+            name='token_concrete_obtain_sync',
+        ),
+        path(
+            'token-concrete-obtain-async/',
+            obtain.ConcreteObtainTokenAsyncController.as_view(),
+            name='token_concrete_obtain_async',
+        ),
+        path(
             'token-custom-sync-auth/',
             obtain.ControllerCustomTokenSync.as_view(),
             name='token_custom_sync_auth',